feat:メモアプリclass化
