+++ src/public/memoList.php
@@ -1,25 +1,12 @@
 <?php
-$pageDepogit = new PageDEpogit();
-$memos = $pageDepogit->pagesDepo();
-echo json_encode($memos);
+require_once __DIR__ . '/../vendor/autoload.php';
+use App\PagesData;
 
-class PageDepogit {
-  private $pdo; 
-  public function __construct() 
-  {
-    $user = 'root';
-    $password = 'password';
-    $pdo = new PDO('mysql:host=mysql; dbname=memo; charset=utf8', $user, $password);
+$pageDepogit = new PagesData();
+$pages = $pageDepogit->fetchPages();
+echo json_encode($pages);
 
-  }
-  public function pagesDepo (){
-    $sql = 'select * from pages';
-    $statement = $pdo->prepare($sql);
-    $statement->execute();
-    $memos = $statement->fetchAll(PDO::FETCH_ASSOC);
-    return $memos;
-  }
-}
+
