+++ src/public/store.php
@@ -1,24 +1,7 @@
 <?php
+require_once __DIR__ . '/../vendor/autoload.php';
+use App\PagesData;
 
-class PageCollect {
-  private $pdo; 
-
-    public function __construct() 
-    {
-      $dbUserName = "root";
-      $dbPassword = "password";
-      $this->pdo = new PDO("mysql:host=mysql; dbname=memo; charset=utf8", $dbUserName, $dbPassword);
-    }
-    public function createPage ($title, $content):void
-    {
-      $sql = 'INSERT INTO `pages`(`title`, `content`) VALUES(:title, :content)';
-      $statement = $this->pdo->prepare($sql);
-      $statement->bindValue(':title', $title, PDO::PARAM_STR);
-      $statement->bindValue(':content', $content, PDO::PARAM_STR);
-      $statement->execute();
-
-    }
-}
 
 
 $content = filter_input(INPUT_POST, 'content');
@@ -28,7 +11,7 @@ $title = filter_input(INPUT_POST, 'title');
 
 // [解説！]ガード節になっている
 if (!empty($title) && !empty($content)) {
-  $pageCollect = new PageCollect();
+  $pageCollect = new PagesData();
   $pageCollect->createPage($title,$content);
   // [解説！]リダイレクト処理
   header('Location: ./index.php');
