+++ src/public/update.php
@@ -1,26 +1,8 @@
 <?php
-class PageDateUpdate 
-{
-  private $pdo; 
-    public function __construct() 
-    {
-      $dbUserName = 'root';
-      $dbPassword = 'password';
-      $this->pdo = new PDO(
-      'mysql:host=mysql; dbname=memo; charset=utf8',
-      $dbUserName,
-      $dbPassword
-      );
-    }
-    public function updatePage ($id, $title, $content) {
-      $sql = 'UPDATE pages SET title=:title, content=:content WHERE id = :id';
-      $statement = $this->pdo->prepare($sql);
-      $statement->bindValue(':id', $id, PDO::PARAM_INT);
-      $statement->bindValue(':title', $title, PDO::PARAM_STR);
-      $statement->bindValue(':content', $content, PDO::PARAM_STR);
-      $statement->execute(); 
-    }
-}
+require_once __DIR__ . '/../vendor/autoload.php';
+use App\PagesData;
+
+
 
 
 $id = filter_input(INPUT_POST, 'id');
@@ -28,7 +10,7 @@ $title = filter_input(INPUT_POST, 'title');
 $content = filter_input(INPUT_POST, 'content');
 
 if (!empty($title) && !empty($content)) {
-    $pageDateUpdate = new PageDateUpdate();
+    $pageDateUpdate = new PagesData();
     $pageDateUpdate->updatePage($id,$title,$content);
 
     header('Location: ./index.php');
